refactor(core): extract catalogs (providers, models, mcp_aliases) to nika-core
Move static catalogs to nika-core/src/catalogs/:
- providers.rs: 19 known providers (LLM, MCP, Local)
- models.rs: 15 known models (types, architecture, quantization)
- mcp_aliases.rs: 48 MCP server short-name aliases

nika/src/core/models.rs kept as thin wrapper for runtime-only
functions (resolve_model, detect_available_ram_gb).

Tests: nika-core 53 passed, nika 6040 passed (41 catalog tests moved).
